requirements.txt
cmake
face_recognition
numpy 
opencv-python
# opencv-python-headless
mediapipe-rpi4
gpiozero
picamera